Moved DI components and finished out XML comments
